Conflicts:
	build/UmbracoVersion.txt
	src/Umbraco.Core/Configuration/UmbracoVersion.cs
	src/Umbraco.Core/Persistence/Repositories/TemplateRepository.cs
	src/Umbraco.Web.UI/Umbraco.Web.UI.csproj

+++ src/UmbracoExamine/LocalStorage/LocalTempStorageIndexer.cs
@@ -1,7 +1,9 @@
-﻿using System.Collections.Specialized;
+﻿using System;
+using System.Collections.Specialized;
 using System.IO;
 using System.Linq;
 using System.Web;
+using Examine.LuceneEngine;
 using Lucene.Net.Analysis;
 using Lucene.Net.Index;
 using Lucene.Net.Store;
@@ -18,27 +20,43 @@ namespace UmbracoExamine.LocalStorage
         public Lucene.Net.Store.Directory LuceneDirectory { get; private set; }
         private readonly object _locker = new object();
         public SnapshotDeletionPolicy Snapshotter { get; private set; }
-        
+
         public LocalTempStorageIndexer()
         {
             IndexDeletionPolicy policy = new KeepOnlyLastCommitDeletionPolicy();
             Snapshotter = new SnapshotDeletionPolicy(policy);
         }
 
-        public void Initialize(NameValueCollection config, string configuredPath, Lucene.Net.Store.Directory baseLuceneDirectory, Analyzer analyzer)
+        public void Initialize(NameValueCollection config, string configuredPath, Lucene.Net.Store.Directory baseLuceneDirectory, Analyzer analyzer, LocalStorageType localStorageType)
         {
             var codegenPath = HttpRuntime.CodegenDir;
 
             _tempPath = Path.Combine(codegenPath, configuredPath.TrimStart('~', '/').Replace("/", "\\"));
 
-            var success = InitializeLocalIndexAndDirectory(baseLuceneDirectory, analyzer, configuredPath);
+            switch (localStorageType)
+            {
+                case LocalStorageType.Sync:
+                    var success = InitializeLocalIndexAndDirectory(baseLuceneDirectory, analyzer, configuredPath);
 
-            //create the custom lucene directory which will keep the main and temp FS's in sync
-            LuceneDirectory = LocalTempStorageDirectoryTracker.Current.GetDirectory(
-                new DirectoryInfo(_tempPath),
-                baseLuceneDirectory,
-                //flag to disable the mirrored folder if not successful
-                success == false);
+                    //create the custom lucene directory which will keep the main and temp FS's in sync
+                    LuceneDirectory = LocalTempStorageDirectoryTracker.Current.GetDirectory(
+                        new DirectoryInfo(_tempPath),
+                        baseLuceneDirectory,
+                        //flag to disable the mirrored folder if not successful
+                        success == false);
+                    break;
+                case LocalStorageType.LocalOnly:
+                    if (Directory.Exists(_tempPath) == false)
+                    {
+                        Directory.CreateDirectory(_tempPath);
+                    }
+
+                    LuceneDirectory = DirectoryTracker.Current.GetDirectory(new DirectoryInfo(_tempPath));
+                    break;
+                default:
+                    throw new ArgumentOutOfRangeException("localStorageType");
+            }
+
         }
 
         private bool InitializeLocalIndexAndDirectory(Lucene.Net.Store.Directory baseLuceneDirectory, Analyzer analyzer, string configuredPath)
@@ -50,7 +68,8 @@ namespace UmbracoExamine.LocalStorage
                     Directory.CreateDirectory(_tempPath);
                 }
 
-                //copy index
+                //copy index if it exists, don't do anything if it's not there
+                if (IndexReader.IndexExists(baseLuceneDirectory) == false) return true;
 
                 using (new IndexWriter(
                     //read from the underlying/default directory, not the temp codegen dir
@@ -64,7 +83,13 @@ namespace UmbracoExamine.LocalStorage
                         var basePath = IOHelper.MapPath(configuredPath);
 
                         var commit = Snapshotter.Snapshot();
-                        var allSnapshotFiles = commit.GetFileNames().Concat(new[] { commit.GetSegmentsFileName() })
+                        var allSnapshotFiles = commit.GetFileNames()
+                            .Concat(new[]
+                            {
+                                commit.GetSegmentsFileName(), 
+                                //we need to manually include the segments.gen file
+                                "segments.gen"
+                            })
                             .Distinct()
                             .ToArray();
 
@@ -75,39 +100,48 @@ namespace UmbracoExamine.LocalStorage
                             .Select(x => x.Name)
                             .Except(allSnapshotFiles);
 
-                        using (var tempDirectory = new SimpleFSDirectory(tempDir))
+                        //using (var tempDirectory = new SimpleFSDirectory(tempDir))
+                        //{
+                        //TODO: We're ignoring if it is locked right now, it shouldn't be unless for some strange reason the 
+                        // last process hasn't fully shut down, in that case we're not going to worry about it.
+
+                        //if (IndexWriter.IsLocked(tempDirectory) == false)
+                        //{
+                        foreach (var file in toRemove)
                         {
-                            if (IndexWriter.IsLocked(tempDirectory) == false)
+                            try
                             {
-                                foreach (var file in toRemove)
-                                {
-                                    try
-                                    {
-                                        File.Delete(Path.Combine(_tempPath, file));
-                                    }
-                                    catch (IOException ex)
-                                    {
-                                        LogHelper.Error<LocalTempStorageIndexer>("Could not delete index file, could not sync from main storage", ex);
-                                        //quit here
-                                        return false;
-                                    }
-                                }
+                                File.Delete(Path.Combine(_tempPath, file));
                             }
-                            else
+                            catch (IOException ex)
                             {
-                                LogHelper.Warn<LocalTempStorageIndexer>("Cannot sync index files from main storage, the index is currently locked");
+                                LogHelper.WarnWithException<LocalTempStorageIndexer>("Could not delete non synced index file file, index sync will continue but old index files will remain - this shouldn't affect indexing/searching operations", ex);
+
+                                //TODO: we're ignoring this, as old files shouldn't affect the index/search operations, lucene files are 'write once'
                                 //quit here
-                                return false;
+                                //return false;
                             }
                         }
+                        //}
+                        //else
+                        //{
+                        //    LogHelper.Warn<LocalTempStorageIndexer>("Cannot sync index files from main storage, the index is currently locked");
+                        //    //quit here
+                        //    return false;
+                        //}
 
                         foreach (var fileName in allSnapshotFiles.Where(f => f.IsNullOrWhiteSpace() == false))
                         {
+                            var destination = Path.Combine(_tempPath, Path.GetFileName(fileName));
+
+                            //don't copy if it's already there, lucene is 'write once' so this file is meant to be there already
+                            if (File.Exists(destination)) continue;
+
                             try
                             {
                                 File.Copy(
                                     Path.Combine(basePath, "Index", fileName),
-                                    Path.Combine(_tempPath, Path.GetFileName(fileName)), true);
+                                    destination);
                             }
                             catch (IOException ex)
                             {
@@ -118,11 +152,17 @@ namespace UmbracoExamine.LocalStorage
                             }
                         }
 
+                        //}
+
+
+
                     }
                     finally
                     {
                         Snapshotter.Release();
                     }
+
+
                 }
 
                 return true;
